repo name is changed.
package object is added instead of utils and io helpers.
some magic strings are handled.
